se agrega comprobación de usuarios duplicados

+++ RF.php
@@ -2,68 +2,74 @@
 
 require("conexion.php");
 
-$con= conectar_bd();
-//comprobar que se envio un formulario por POST desde carga_datos
-if(isset($_POST["envio"])){
+$con = conectar_bd();
 
+// Comprobar que se envió un formulario por POST desde carga_datos
+if (isset($_POST["envio"])) {
 
-    $nombre= $_POST["nombre"];
-    $apellido= $_POST["apellido"];
-    $email= $_POST["email"];
-    $contrasenia= $_POST["pass"];
+    $nombre = $_POST["nombre"];
+    $apellido = $_POST["apellido"];
+    $email = $_POST["email"];
+    $contrasenia = $_POST["pass"];
    
-    insertar_datos($con, $nombre, $apellido, $email, $contrasenia);
+    // Consultar si el usuario ya existe
+    $existe_usr = consultar_existe_usr($con, $email);
+
+    // Insertar datos si el usuario no existe
+    insertar_datos($con, $nombre, $apellido, $email, $contrasenia, $existe_usr);
 }
 
+function consultar_existe_usr($con, $email) {
 
-function insertar_datos($con, $nombre, $apellido, $email, $contrasenia){
+    $email = mysqli_real_escape_string($con, $email); // Escapar los campos  para evitar inyección SQL
+    $consulta_existe_usr = "SELECT email FROM usuarios WHERE email = '$email'";
+    $resultado_existe_usr = mysqli_query($con, $consulta_existe_usr);
 
+    if (mysqli_num_rows($resultado_existe_usr) > 0) {
+        return true;
+    } else {
+        return false;
+    }
+}
 
+function insertar_datos($con, $nombre, $apellido, $email, $contrasenia, $existe_usr) {
 
-    $nombreCompleto= $nombre . ' '. $apellido;
+    if ($existe_usr == false) {
+        $nombreCompleto = $nombre . ' ' . $apellido;
+        $email = mysqli_real_escape_string($con, $email);
+        $contrasenia = mysqli_real_escape_string($con, $contrasenia);
 
-    $consulta_insertar= "INSERT INTO usuarios(nombrecompleto, email, pass )  VALUES( '$nombreCompleto', '$email', '$contrasenia') ";
+        $consulta_insertar = "INSERT INTO usuarios (nombrecompleto, email, pass) VALUES ('$nombreCompleto', '$email', '$contrasenia')";
 
-    if (mysqli_query($con, $consulta_insertar)) {
-        
-
-        consultar_datos($con);
-
-
-  } else {
-        echo "Error: " . $consulta_insertar . "<br>" . mysqli_error($con);
-  }
-
+        if (mysqli_query($con, $consulta_insertar)) {
+            $salida = consultar_datos($con);
+            echo $salida;
+        } else {
+            echo "Error: " . $consulta_insertar . "<br>" . mysqli_error($con);
+        }
+    } else {
+        echo "El usuario ya existe.";
+    }
 }
 
 function consultar_datos($con) {
-
-
     $consulta = "SELECT * FROM usuarios";
-    
     $resultado = mysqli_query($con, $consulta);
-   
+
     // Inicializo una variable para guardar los resultados
     $salida = "";
-   
-    //si se recupera algun registro de la consulta
-    if (mysqli_num_rows($resultado) > 0) {
 
-        //mientras haya registros..
+    // Si se recupera algún registro de la consulta
+    if (mysqli_num_rows($resultado) > 0) {
+        // Mientras haya registros
         while ($fila = mysqli_fetch_assoc($resultado)) {
             $salida .= "id: " . $fila["id_user"] . " - Nombre: " . $fila["nombrecompleto"] . " - Email: " . $fila["email"] . "<br> <hr>";
         }
     } else {
-        $salida = "Sin datos ";
+        $salida = "Sin datos";
     }
 
     return $salida;
 }
 
-
-//puedo darle un formato a la salida
-
-$salida= consultar_datos($con);
-mysqli_close($con);   
-
-echo $salida;
+mysqli_close($con);
